+++ lint.py
@@ -19,9 +19,12 @@ class Linter:
             exit(1)
 
 
-linter = Linter(flake8=("flake8", "--exclude",
-                        "spezspellz/migrations,__init__.py",
-                        "--max-line-length", "256", "spezspellz"),
-                mypy=("mypy", "--ignore-missing-imports", "spezspellz"),
-                ruff=(("ruff", "check", "spezspellz")))
+linter = Linter(
+    flake8=(
+        "flake8", "--exclude", "spezspellz/migrations", "--max-line-length",
+        "256", "spezspellz"
+    ),
+    mypy=("mypy", "--ignore-missing-imports", "spezspellz"),
+    ruff=(("ruff", "check", "spezspellz"))
+)
 linter.run()
