+++ spezspellz/__init__.py
@@ -1,0 +1,1 @@
+"""Makes spezspellz a module."""
