+++ spezspellz/views/register_view.py
@@ -1,0 +1,12 @@
+"""Implements the register page."""
+from django.urls import reverse_lazy
+from django.contrib.auth.forms import UserCreationForm
+from django.views.generic import CreateView
+
+
+class RegisterView(CreateView):
+    """View for registration/signup."""
+
+    form_class = UserCreationForm
+    success_url = reverse_lazy("login")
+    template_name = "registration/register.html"
